+++ lib/src/wrapper/youth_video_error_data.dart
@@ -1,0 +1,6 @@
+class YouthVideoErrorData {
+  final int code;
+  final String message;
+
+  YouthVideoErrorData(this.code, this.message);
+}