+++ lib/src/wrapper/youth_video_warning_data.dart
@@ -1,0 +1,6 @@
+class YouthVideoWarningData {
+  final int code;
+  final String message;
+
+  YouthVideoWarningData(this.code, this.message);
+}