+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## 0.1.0
-// initial version of the biomarkers SDK
+- initial version of the biomarkers SDK
+## 0.2.0
+-  Implemented Binah video service v1
+-  Warning and Error callbacks handled
+in appropriate format for the client side